+++ releasenotes.md
@@ -11,6 +11,7 @@
 - Macros with "const" declarations were compile time folded despite needing the declaration.
 - Vectors created through casts in globals would miscompile, e.g. `Foo global = { .v = (int[<3>])8 }`.
 - Support $$bswap for all power of two integers, including single bytes (in which case it is a no-op).
+- Fix source locations for byte literal values continued on a new line.
 
 ### Stdlib changes
 - Added `std::gfx` module.
@@ -47,7 +48,6 @@
 - Add `@noredzone` attribute.
 - Add `dso_local` attribute on ELF/COFF.
 - Let LLVM build the biggest modules first.
-- Fix source locations for byte literal values continued on a new line.
 - Support for iOS.
 - Allow `-` in `c3c init some-project`.
 - `--keep-obj` added, to prevent object files from being deleted after building/linking.